like funcitonality with ajax

+++ restaurant/views.py
@@ -23,19 +23,21 @@ def menu_cat(request, menu_cat_name):
 
 	for food_menu in food_menu_list:
 		curr_like_list = LikeFoodMenu.objects.filter(food_menu = food_menu)
-		like_status = ''
+		is_like = ''
+		num_other_like = curr_like_list.count()
 
 		if is_signed_in and curr_like_list.filter(customer = request.user):
-			like_status = True
+			is_like = True
+			num_other_like = curr_like_list.count() - 1
 		elif is_signed_in and not curr_like_list.filter(customer = request.user):
-			like_status = False
+			is_like = False
 		else:
-			like_status = False
+			is_like = False
 
 		if food_menu.food_cat not in menu_dict:
-			menu_dict[food_menu.food_cat] = [(food_menu, curr_like_list.count(), like_status, request.user.id)]
+			menu_dict[food_menu.food_cat] = [(food_menu, num_other_like, is_like, request.user.id)]
 		else:
-			menu_dict[food_menu.food_cat].append((food_menu, curr_like_list.count(), like_status, request.user.id))
+			menu_dict[food_menu.food_cat].append((food_menu, num_other_like, is_like, request.user.id))
 
 	context = {'menu_cat': menu_cat, 'menu_dict' : menu_dict}
 	return render(request, 'restaurant/menu_cat.html', context)
@@ -77,14 +79,16 @@ def album(request, album_name):
 	return render(request, 'restaurant/album.html', context)
 
 def food_menu_like(request, food_menu_id):
-	if request.user.is_authenticated() and food_menu_id:
-		curr_food_menu = FoodMenu.objects.get(id = int(food_menu_id))
-		(obj, created) = LikeFoodMenu.objects.get_or_create(customer = request.user, food_menu = curr_food_menu)
-		if created == False:
-			obj.delete()
-		return HttpResponseRedirect('/restaurant/menu/%s' % curr_food_menu.menu_cat.name) 
-	else:
-		return HttpResponseRedirect('/restaurant/') 
+	if request.method == 'POST':
+		num_other_like = request.POST['num_other_like']
+		if request.user.is_authenticated() and food_menu_id:
+			curr_food_menu = FoodMenu.objects.get(id = int(food_menu_id))
+			(obj, created) = LikeFoodMenu.objects.get_or_create(customer = request.user, food_menu = curr_food_menu)
+			if created == False:
+				obj.delete()
+				return render(request, 'ajax_like_count.html' , {'is_like' : False, 'num_other_like' : int(num_other_like)})
+			return render(request, 'ajax_like_count.html' , {'is_like' : True, 'num_other_like' : int(num_other_like)})
+	return HttpResponseRedirect('/restaurant/')
